Create Registration Generator.py
I am giving you the requirements and the registration formation :

 I want to customize the registration ID. Here is:

registration contains 10 digit .

First digit is 3 which will be same for everyone. 3 = donotes Savar Science Society Second digit : 01 for When student select SCPSC 02 for When student select JUSC 03 for When student select BPATC 04 for when student select GBS 05 For when students select SCLS

Third Digit : 05 for class 5 06 for class 6 07 for class 7 08 for class 8 09 for class 9 10 for class 10

Fourth digit: 1 for When student will select Science 2 for When student will select Math 3 for When student will select Physics

Note : The Science and Math will be available when the student is in class 5 to 8

If student select 9 to 10 then , He/She will be able to select Math or Physics for exam

Fifth digit is Gender selection: 1 for Male 2 For Female

Sixth Digit is : 001 to 999

Now, for school , class and gender wise the registration number will be new

Suppose,

A student select school - 01 Then select class - 07 Then select Math - 2 And select Gender - 01(Male)

Then his number will be : 3010721001

if the student is female for that school ( 01 ) Then the number will be : 3010722001

Again , if the student is from another school like : 03 - BPATC class: 08 Subject : Science Gender : Male (1) and if the student is Male his Registration number will be : 3030811001 If the student is in same class , same subject But Female , Then Registration Number will be : 3030812001

I think you understand. Now make this registration algorithm to provide registration number to the students.